PostService/backend/requirements.txt
asyncio==4.0.0
sqlalchemy==2.0.43
fastapi[standard]==0.116.1
pydantic-settings==2.10.1
asyncpg==0.30.0
pytest==8.4.1